+++ src/c_helper.h
@@ -1,6 +1,9 @@
 #ifndef C_HELPER_H_
 #define C_HELPER_H_
 
+#include <iomanip>
+#include <limits>
+
 /*
 reference: http://www.iiug.org/software/archive/cfuncts
 this is an old repository of very nice functions check it :)
@@ -15,5 +18,12 @@ char *strremchar(char *str, int c, char *target);
 void strcpy(const char *from, char* to);
 
 #define MAX(A,B) ((A)>(B) ? (A) : (B))
+
+//
+// ----------------------------------------------------------------------------------------------------
+//
+
+int nextpoweroftwo(float x);
+
 
 #endif // C_HELPER_H_
