+++ src/radamn.cc
@@ -1,7 +1,6 @@
 #include "radamn.h"
 
 #include "prerequisites.h"
-#include "SDL_helper.h"
 #include "v8_helper.h"
 
 #include "radamn_image.h"
@@ -10,7 +9,7 @@
 
 #include <SDL_version.h>
 #include <SDL_ttf.h>
-#include "radamn_gl.h"
+#include "gl.h"
 #include <node.h>
 #include <v8.h>
 
@@ -237,20 +236,19 @@ static v8::Handle<v8::Value> radamn::createWindow(const v8::Arguments& args) {
     }
 
     SDL_GL_SetAttribute(SDL_GL_DOUBLEBUFFER,    1);
-    SDL_GL_SetAttribute(SDL_GL_STENCIL_SIZE,    0);
 
     SDL_GL_SetAttribute(SDL_GL_RED_SIZE,        8);
     SDL_GL_SetAttribute(SDL_GL_GREEN_SIZE,      8);
     SDL_GL_SetAttribute(SDL_GL_BLUE_SIZE,       8);
     SDL_GL_SetAttribute(SDL_GL_ALPHA_SIZE,      8);
 
-    SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE,      32);
+    SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE,      16);
     SDL_GL_SetAttribute(SDL_GL_BUFFER_SIZE,     32);
 
-    SDL_GL_SetAttribute(SDL_GL_ACCUM_RED_SIZE,    0);
-    SDL_GL_SetAttribute(SDL_GL_ACCUM_GREEN_SIZE,  0);
-    SDL_GL_SetAttribute(SDL_GL_ACCUM_BLUE_SIZE,   0);
-    SDL_GL_SetAttribute(SDL_GL_ACCUM_ALPHA_SIZE,  0);
+    SDL_GL_SetAttribute(SDL_GL_ACCUM_RED_SIZE,    8);
+    SDL_GL_SetAttribute(SDL_GL_ACCUM_GREEN_SIZE,  8);
+    SDL_GL_SetAttribute(SDL_GL_ACCUM_BLUE_SIZE,   8);
+    SDL_GL_SetAttribute(SDL_GL_ACCUM_ALPHA_SIZE,  8);
 
     SDL_GL_SetAttribute(SDL_GL_MULTISAMPLEBUFFERS,  1);
     SDL_GL_SetAttribute(SDL_GL_MULTISAMPLESAMPLES,  2);
